requirements.txt
flask==3.0.0
python-dotenv==1.0.0
supabase==1.2.0
httpx>=0.24.1
python-dateutil==2.8.2 